test_setup.py
import torch
from transformers import AutoTokenizer
from sentence_transformers import SentenceTransformer
from langchain.chains import LLMChain  # Updated import
import faiss
from peft import LoraConfig

print("All libraries imported successfully!")