requirements.txt
torch>=2.0.0
transformers>=4.44.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
peft>=0.4.0
pandas>=1.5.0
numpy>=1.21.0
huggingface_hub>=0.16.0
